Refs #4964. Move CLI classes to utils.
